move updatequeue to telegram.utils
